requirements.txt
alembic==1.11.1
async-timeout==4.0.3
blinker==1.6.2
boto3==1.34.44
botocore==1.34.44
cachelib==0.10.2
certifi==2022.12.7
click==8.1.4
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Assets==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
redis==5.0.1
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.7.1
urllib3==2.0.7
webassets==2.0
Werkzeug==2.3.3
WTForms==3.0.1
